Time cooldowns
You are now allowed to say "o" once every 30 seconds

+++ o_bot.js
@@ -11,33 +11,45 @@ hash["oo"] = "o";
 
 const whitelist = ['o-o'];
 
+const timerCooldown = 30 * 1000; // ms
+var timers = {};
+
 client.on('ready', () => {
   console.log(`Logged in as ${client.user.tag}!`);
 });
 
 client.on('message', msg => {
+  check(msg);
+});
 
+client.on('messageUpdate', function (oldMsg, msg) {
+  check(msg);
+});
+
+// Checks the message and determines whether or not it needs to be deleted
+function check(msg) {
   // Debug
   // console.log('message: ' + msg.content + '  isAnO:' + msg.content.isAnO());
 
-  // If the message is just an "o" then delete it
+  // If the message is an "o"
   if (msg.content.isAnO()) {
-    msg.delete();
+    
+    // If the author has a timer on them already
+    if (timers[msg.author.id]) {
+      // If it's been less than the timer cooldown since the last "o"
+      if (Date.now() - timers[msg.author.id] < timerCooldown) {
+        // Delete the message
+        msg.delete();
+      }
+    }
+    
+    // Reset the timer for every "o"
+    timers[msg.author.id] = Date.now();
   }
-});
+}
 
-client.on('messageUpdate', function(oldMsg, msg) {
-
-  // Debug
-  // console.log('message: ' + msg.content + '  isAnO:' + msg.content.isAnO());
-
-  // If the message is just an "o" then delete it
-  if (msg.content.isAnO()) {
-    msg.delete();
-  }
-});
-
-
+// To remove a timer
+// delete timers[msg.author.id];
 
 
 // String functions
@@ -51,12 +63,12 @@ String.prototype.isAnO = function () {
 
   // If the string contains "o" and no other real alphabet characters that aren't o
   if (text.contains('o') && !/^([A-Za-z0-9]+)$/.test(text.replaceAll('o', ''))) {
-    
+
     // Last chance before confirming it's an "o". If the string is equal to anything in the whitelist, let it through
     for (i in whitelist) {
       if (text == whitelist[i]) return false;
     }
-    
+
     return true;
   } else {
     return false;
